# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flow ERP - Odoo Supply Chain Read-Only API",
    "description": "Secure, read-only bridge connecting ChatGPT to OneScreen's Colombia Odoo ERP.",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Root",
        "description": "Service status and readiness check.",
        "operationId": "root",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health Check",
        "description": "Health check for external monitoring.",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/activities/overdue": {
      "get": {
        "tags": [
          "Supply Chain Activities"
        ],
        "summary": "Get overdue activities",
        "description": "Fetches Supply Chain activities where the deadline is in the past.",
        "operationId": "get_overdue_activities",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/pos/open": {
      "get": {
        "tags": [
          "Purchase Orders"
        ],
        "summary": "Get open purchase orders",
        "description": "Fetches open Colombia Purchase Orders (draft, sent, purchase).",
        "operationId": "get_open_purchase_orders",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/pos/{po_name}": {
      "get": {
        "tags": [
          "Purchase Orders"
        ],
        "summary": "Get purchase order details",
        "description": "Fetches exact details of a specific Purchase Order by name (e.g., P00012), including line items, quantities, and prices.",
        "operationId": "get_purchase_order_details",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "po_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rmas/open": {
      "get": {
        "tags": [
          "RMAs"
        ],
        "summary": "Get open RMAs",
        "description": "Fetches open RMAs from the rma_from_forms module.",
        "operationId": "get_open_rmas",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/rmas/{rma_name}": {
      "get": {
        "tags": [
          "RMAs"
        ],
        "summary": "Get RMA details",
        "description": "Fetches exact details of a specific RMA by name.",
        "operationId": "get_rma_details",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rma_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rma Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer token authentication. Enter your GPT_SECRET_TOKEN.",
        "scheme": "bearer"
      }
    }
  }
}